+++ testutil/Fixture/template/partials/outer.html.php
@@ -1,0 +1,10 @@
+<?php
+
+use SubstancePHP\HTTP\Renderer\HtmlRenderer;
+
+/**
+ * @var HtmlRenderer $this
+ * @var string $word
+ */
+?>
+[<?= $this->partial('inner', ['word' => $word]) ?>]